rename Sign.vue to SignTop.vue
